requirements.txt
mongomock
pyyaml
pytest-asyncio
pytest>=2.5.2
